pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.7",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "sqlalchemy>=2.0.48",
    "tqdm>=4.67.3",
    "google-api-core>=2.25.0",
    "google-cloud-storage>=2.19.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pgcli>=4.4.0",
]
